allow ActiveState's Win32.pm to load XS in miniperl.t
And fixing the stacktrace reveals that Win32.pm is the lone module that tries to load XS under ActivePerl because their Config.pm needs it to resolve some paths. As such we store the original bootstrap and allow Win32.pm to pass through to it.

With this change Win32.pm is the only module that is allowed to do XS loading, so i think the spirit of the test is still done justice. I'm not too sure as to how useful this test is under these circumstances, but it should not generate false positives nor false negatives.